+++ spec/fixtures/move_action_with_implicit_default_strategy.rb
@@ -1,0 +1,5 @@
+require 'strategic'
+
+class MoveActionWithImplicitDefaultStrategy
+  include Strategic
+end