+++ spec/fixtures/move_action_with_implicit_default_strategy/default_strategy.rb
@@ -1,0 +1,9 @@
+class MoveActionWithImplicitDefaultStrategy
+  class DefaultStrategy
+    include Strategic::Strategy
+    
+    def move
+      context.position += 10
+    end
+  end
+end